backend/pyproject.toml
[project]
name = "smart-lock-backend"
version = "0.1.0"
description = "Smart Lock IoT Backend with Face Recognition, RFID, and Keypad"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "pyserial>=3.5",
    "python-multipart>=0.0.19",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "numpy>=2.2.1",
    "opencv-python>=4.10.0.84",
]
